requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
openpyxl==3.1.5
python-multipart==0.0.9
gunicorn==25.1.0
pdfplumber==0.11.4
PyMuPDF==1.24.14
pytesseract==0.3.13
Pillow==11.0.0
python-docx==1.1.2
